Added RR interval, pulse arrival time, and SQI.
Note that the SQI function in neurokit2 is meant for ECG signals. Presumably if an ECG signal is good, the PPG signal is good too. But data collected from the back of the device does not have ECG data, so we'll need another way to assess the quality of those log files.